- GL clip planes should never be a negative number. Move to use of size_t   and avoid casting.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@31387 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csplugincommon/opengl/shaderplugin.h
@@ -115,8 +115,8 @@ namespace CS
       /// Helper class for user defined OpenGL clip planes.
       class CS_CSPLUGINCOMMON_GL_EXPORT ClipPlanes
       {
-        int maxPlanes;
-        uint32 currentPlanes;
+        size_t maxPlanes;
+        size_t currentPlanes;
         
         CS::ShaderVarStringID svObjectToWorldInv;
         CS::ShaderVarStringID svWorldToCamera;
@@ -154,7 +154,7 @@ namespace CS
          * Enable the \a n th OpenGL clip plane without setting it. It is up 
          * to the caller to call glClipPlane().
          */
-        bool EnableClipPlane (int n);
+        bool EnableClipPlane (size_t n);
         /**
          * Enable the next available OpenGL clip plane without setting it. 
          * It is up  to the caller to call glClipPlane().
